# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Pre-Autorización Quirúrgica",
    "description": "API del agente de pre-autorización quirúrgica en tiempo real. Recibe informes médicos + pólizas, valida cobertura/carencia/docs, y emite pre-aprobación, solicitud de docs faltantes o escalamiento al auditor médico humano. NUNCA auto-rechaza por diseño.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/patients": {
      "get": {
        "tags": [
          "patients"
        ],
        "summary": "List or search patients",
        "operationId": "list_patients_api_v1_patients_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "dni",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filtra por DNI exacto (devuelve 0 ó 1).",
              "title": "Dni"
            },
            "description": "Filtra por DNI exacto (devuelve 0 ó 1)."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PatientOut"
                  },
                  "title": "Response List Patients Api V1 Patients Get"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/patients/{patient_id}": {
      "get": {
        "tags": [
          "patients"
        ],
        "summary": "Get a single patient by id",
        "operationId": "get_patient_api_v1_patients__patient_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "patient_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Patient Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientOut"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token."
          },
          "404": {
            "description": "Patient not found."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PatientOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "dni": {
            "type": "string",
            "title": "Dni"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "dob": {
            "type": "string",
            "format": "date",
            "title": "Dob"
          },
          "sex": {
            "type": "string",
            "title": "Sex"
          }
        },
        "type": "object",
        "required": [
          "id",
          "dni",
          "name",
          "dob",
          "sex"
        ],
        "title": "PatientOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}